https://drive.google.com/file/d/1IVbEViuP8ISdgfA_gFh8xjyeapLX4c2K/view
Multiphase flows take place in nuclear geothermal power plants, food processing industries, and many others. One of the basic examples is Rayleigh Taylor Instability. The Rayleigh Taylor Instability is a hydrodynamic instability that arises when a high-density fluid is placed over a low-density fluid in a gravitational field. When the density interface is disturbed, the hydrostatic pressure is generated, so the heavier fluid moves downwards and mixes with the lighter one. The Rayleigh Taylor instability is the instability of an interface between two fluids of different densities which occurs when the lighter fluid is pushing the heavier fluid in a gravitational field.
I have attached the drive link of the simulation and mesh image with the 20 thousand cell.